@@ -2,73 +2,37 @@
 
 cmake_minimum_required(VERSION 3.20 FATAL_ERROR)
 
-# TODO: Debug
-include(CMakePrintHelpers)
-include(CheckLanguage)
-
 set(EXE_NAME GenArt)
 
 # CUDA
 project(${EXE_NAME} LANGUAGES CXX CUDA)
 enable_language(CUDA)
-# set_property(TARGET ${EXE_NAME} PROPERTY CUDA_ARCHITECTURES 86)
 
-# TODO: Debug
-check_language(CUDA)
-cmake_print_variables(CMAKE_CUDA_COMPILER)
-
-set(CMAKE_CXX_STANDARD 11 CACHE STRING "" FORCE)
-set_property(GLOBAL PROPERTY USE_FOLDERS ON)
-
+set(CMAKE_CXX_STANDARD 20)
 set(PROJECT_ROOT_DIR "${CMAKE_CURRENT_SOURCE_DIR}")
 
-# warning level, all warnings as errors, build in parallel, disable security warnings, latest language support
-# add_compile_options(/W3 /WX /MP /sdl- /std:c++latest)
-add_compile_options($<$<COMPILE_LANGUAGE:CXX>:/W3>)
-add_compile_options($<$<COMPILE_LANGUAGE:CXX>:/WX>)
-add_compile_options($<$<COMPILE_LANGUAGE:CXX>:/MP>)
-add_compile_options($<$<COMPILE_LANGUAGE:CXX>:/sdl->)
-add_compile_options($<$<COMPILE_LANGUAGE:CXX>:/std:c++latest>)
-add_compile_options($<$<COMPILE_LANGUAGE:CUDA>:-std>)
-add_compile_options($<$<COMPILE_LANGUAGE:CUDA>:c++17>)
-
-set(__ARCH "x64")
-
-# GLEW
+# GLEW - setenv GLEW_HOME C:/Users/davemc/source/repos/Goodies/glew-2.2.0
+set(CMAKE_PREFIX_PATH $ENV{GLEW_HOME})
+set(CMAKE_LIBRARY_PATH $ENV{GLEW_HOME}/lib/Release/x64)
 set(GLEW_USE_STATIC_LIBS TRUE)
-set(CMAKE_PREFIX_PATH "../Goodies/glew-2.2.0")
-set(CMAKE_LIBRARY_PATH "../Goodies/glew-2.2.0/lib/Release/x64")
 
 find_package(GLEW REQUIRED)
-# TODO: Debug
-cmake_print_variables(GLEW_INCLUDE_DIRS)
-cmake_print_variables(GLEW_STATIC_LIBRARIES)
 
 add_definitions(-DGLEW_STATIC)
 include_directories(${GLEW_INCLUDE_DIRS})
 link_libraries(${GLEW_STATIC_LIBRARIES})
 
-# FreeGLUT
-set(GLUT_INCLUDE_DIR "../Goodies/freeglut-3.2.2/include")
-set(GLUT_LIBRARIES "optimized;C:/Users/davemc/source/repos/Goodies/freeglut-3.2.2/lib/Release/freeglut_static.lib;debug;C:/Users/davemc/source/repos/Goodies/freeglut-3.2.2/lib/Debug/freeglut_staticd.lib")
+# FreeGLUT - setenv GLUT_HOME C:/Users/davemc/source/repos/Goodies/freeglut-3.2.2
+set(GLUT_INCLUDE_DIR $ENV{GLUT_HOME}/include)
+set(GLUT_LIBRARIES "optimized;$ENV{GLUT_HOME}/lib/Release/freeglut_static.lib;debug;$ENV{GLUT_HOME}/lib/Debug/freeglut_staticd.lib")
 
 add_definitions(-DFREEGLUT_STATIC)
 include_directories(${GLUT_INCLUDE_DIR})
 link_libraries(${GLUT_LIBRARIES})
 
-set(SOURCES
-    src/AutoScorer.h
-    src/AutoScorer_kernels.cu
+set(EXPR_SOURCES
     src/BinaryExprSubclasses.cpp
     src/BinaryExprSubclasses.h
-    src/CUDAHelpers.cpp
-    src/CUDAHelpers.h
-    src/CUDARender.h
-    src/Counters.cpp
-    src/Counters.h
-    src/Evaluator.h
-    src/Evolver.cpp
-    src/Evolver.h
     src/Expr.cpp
     src/Expr.h
     src/ExprHelpers.cpp
