@@ -1,4 +1,5 @@
 #pragma once
 
 extern void Test();
-extern void TestExpressions(const char* exrFName = NULL);
+extern void TestExpressions(const char* exrFName = nullptr);
+extern float randVal(); // Used by all the testing code
