@@ -2,6 +2,7 @@
 
 #include "BinaryExprSubclasses.h"
 #include "ExprImplementations.h"
+#include "Math/BinaryRep.h"
 #include "Math/Random.h"
 #include "UnaryExprSubclasses.h"
 #include "Util/Assert.h"
@@ -18,26 +19,11 @@ std::string Var::name = "var";
 std::string Const::fname = "CONSTISBROKEN";
 std::string Var::fname = "VARISBROKEN";
 
-union UFloatInt
-{
-    int i;
-    float f;
-};
+float Const::Eval(const VarVals_t* VV) const { return val; }
 
-float Const::Eval(const VarVals_t* VV) const
-{
-    return val;
-}
+interval Const::Ival(const opInfo& opI, const interval& lv /* = interval() */, const interval& rv /*= interval() */) const { return interval(val); }
 
-interval Const::Ival(const opInfo& opI, const interval& lv /* = interval() */, const interval& rv /*= interval() */) const
-{
-    return interval(val);
-}
-
-Expr* Const::Copy() const
-{
-    return new Const(*this);
-}
+Expr* Const::Copy() const { return new Const(*this); }
 
 std::string Const::Print(int pstyle) const
 {
@@ -49,9 +35,7 @@ std::string Const::Print(int pstyle) const
 int Const::preTokenStream(int* TokenStream, const int max_len) const
 {
     *(TokenStream++) = token;
-    UFloatInt FI;
-    FI.f = val;
-    *(TokenStream++) = FI.i;
+    *(TokenStream++) = floatAsUint(val);
     return 2;
 }
 
@@ -59,9 +43,7 @@ int Const::postTokenStream(int* TokenStream, const int max_len) const
 {
     ASSERT_R(max_len >= 2);
     *(TokenStream++) = token;
-    UFloatInt FI;
-    FI.f = val;
-    *(TokenStream++) = FI.i;
+    *(TokenStream++) = floatAsUint(val);
     return 2;
 }
 
@@ -84,10 +66,7 @@ Expr* Const::PerturbConstants(const float ConstPerturbStDev)
     return this;
 }
 
-bool Const::isequal(const Expr* E) const
-{
-    return typeid(*E) == typeid(Const) && Eval() == ((const Const*)E)->Eval();
-}
+bool Const::isequal(const Expr* E) const { return typeid(*E) == typeid(Const) && Eval() == ((const Const*)E)->Eval(); }
 
 bool Const::isless(const Expr* E) const
 {
@@ -97,10 +76,7 @@ bool Const::isless(const Expr* E) const
         return getToken() < E->getToken();
 }
 
-void Const::UniformRandom()
-{
-    val = DRandf();
-}
+void Const::UniformRandom() { val = frand(); }
 
 Const::Const()
 {
@@ -122,20 +98,11 @@ float Var::Eval(const VarVals_t* VV /*= NULL*/) const
     return VV->vals[VarID];
 }
 
-interval Var::Ival(const opInfo& opI, const interval& lv /* = interval() */, const interval& rv /*= interval() */) const
-{
-    return interval(opI.spans[VarID]);
-}
+interval Var::Ival(const opInfo& opI, const interval& lv /* = interval() */, const interval& rv /*= interval() */) const { return interval(opI.spans[VarID]); }
 
-Expr* Var::Copy() const
-{
-    return new Var(*this);
-}
+Expr* Var::Copy() const { return new Var(*this); }
 
-std::string Var::Print(int pstyle) const
-{
-    return VarName + ((pstyle & PREFIX) ? " " : "");
-}
+std::string Var::Print(int pstyle) const { return VarName + ((pstyle & PREFIX) ? " " : ""); }
 
 int Var::preTokenStream(int* TokenStream, const int max_len) const
 {
@@ -170,10 +137,7 @@ Expr* Var::Mutate(const int prob, const int siz, const float ConstPerturb, const
     return NULL;
 }
 
-bool Var::isequal(const Expr* E) const
-{
-    return typeid(*E) == typeid(Var) && getVarID() == ((const Var*)E)->getVarID();
-}
+bool Var::isequal(const Expr* E) const { return typeid(*E) == typeid(Var) && getVarID() == ((const Var*)E)->getVarID(); }
 
 bool Var::isless(const Expr* E) const
 {
@@ -183,10 +147,7 @@ bool Var::isless(const Expr* E) const
         return getToken() < E->getToken();
 }
 
-Var::Var()
-{
-    count = 0;
-}
+Var::Var() { count = 0; }
 
 Var::Var(const std::string VarName_, int VarID_)
 {
@@ -196,12 +157,6 @@ Var::Var(const std::string VarName_, int VarID_)
     count = 1;
 }
 
-std::string Var::GetVarName() const
-{
-    return VarName;
-}
+std::string Var::GetVarName() const { return VarName; }
 
-size_t Var::getVarID() const
-{
-    return VarID;
-}
+size_t Var::getVarID() const { return VarID; }
