@@ -1,6 +1,3 @@
 #include "Style.h"
 
-Style::Style()
-{
-    m_variability = 0.5f;
-}
+Style::Style() { m_variability = 0.5f; }
